CloudRun/requirements.txt
Flask==3.0.3
gunicorn==23.0.0
Werkzeug==3.0.3
requests>=2.20.0
google-cloud-pubsub>=2.0.0